fix(starthandler): Fix the start handler
STATE:
According the chef documentation (ref: https://docs.chef.io/handlers/),
a start handler could not be loaded with the resource chef_handler:
"A start handler is used to run events at the beginning of a Chef Infra
Client run. A start handler can be loaded at the start of a Chef Infra
Client run by adding the start handler to the start_handlers setting in
the client.rb file or by installing the gem that contains the start
handler by using the chef_gem resource in a recipe in the chef-client
cookbook. (A start handler may not be loaded using the chef_handler
resource.)"
So depending on the 'client.rb' configuration is not a good enough
solution.

PROPOSED SOLUTION:
There are ways with chef to subscribe to certain events with libraries,
so the start handler will be loaded on the 'library_load_complete' (not
all handlers are accessible during a chef run, ref:
https://github.com/chef/chef/issues/4314).
This event is the early enough to stop the chef run if '/etc' is not
clean.

NOTE:
After some quick research I found other start handlers:
- https://onddo.github.io/chef-handler-zookeeper/
- https://github.com/needle-cookbooks/chef-sensu-handler/blob/master/recipes/default.rb
- https://github.com/DataDog/chef-datadog/blob/main/recipes/dd-handler.rb
But all the above solutions have one thing in common, there is no
idempotency due to the 'ruby_block' running during the compilation phase.

Signed-off-by: Jeremy MAURO <jeremy.mauro@gmail.com>